Fix MSVC and Clang-Tidy warnings
- Fixed MSVC Warning C4127 (conditional expression is constant) in model_description_checker.h by using if constexpr.
- Fixed MSVC Warning C4244 (possible loss of data) in zipper.cpp by adding explicit static_cast<uint16_t>.
- Removed unused #include <iterator> from model_checker.cpp to resolve clang-tidy warning.
- Updated .clang-tidy HeaderFilterRegex to match the renamed 'core/' directory.

+++ core/zipper.cpp
@@ -87,9 +87,9 @@ std::vector<ZipFileEntry> Zipper::getEntries() const
 
         ZipFileEntry entry;
         entry.filename = filename.data();
-        entry.compression_method = file_info.compression_method;
-        entry.version_needed = file_info.version_needed;
-        entry.flags = file_info.flag;
+        entry.compression_method = static_cast<uint16_t>(file_info.compression_method);
+        entry.version_needed = static_cast<uint16_t>(file_info.version_needed);
+        entry.flags = static_cast<uint16_t>(file_info.flag);
         entry.compressed_size = file_info.compressed_size;
         entry.uncompressed_size = file_info.uncompressed_size;
         entry.is_encrypted = (file_info.flag & 0x01) != 0;
